requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.4.1
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
autogen==0.3.0
    # via -r requirements.in
azure-core==1.31.0
    # via azure-identity
azure-identity==1.17.1
    # via -r requirements.in
blinker==1.8.2
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==43.0.1
    # via
    #   azure-identity
    #   msal
    #   pyjwt
diskcache==5.6.3
    # via
    #   autogen
    #   pyautogen
distro==1.9.0
    # via openai
docker==7.1.0
    # via
    #   autogen
    #   pyautogen
flaml==2.2.0
    # via
    #   autogen
    #   pyautogen
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jiter==0.5.0
    # via openai
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msal==1.31.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
narwhals==1.8.0
    # via altair
numpy==1.26.4
    # via
    #   autogen
    #   flaml
    #   pandas
    #   pyarrow
    #   pyautogen
    #   pydeck
    #   streamlit
openai==1.45.0
    # via
    #   autogen
    #   pyautogen
packaging==24.1
    # via
    #   altair
    #   autogen
    #   pyautogen
    #   streamlit
pandas==2.2.2
    # via streamlit
pillow==10.4.0
    # via streamlit
portalocker==2.10.1
    # via msal-extensions
protobuf==4.25.4
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pyautogen==0.3.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.9.1
    # via
    #   autogen
    #   openai
    #   pyautogen
pydantic-core==2.23.3
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.9.0
    # via
    #   msal
    #   pyjwt
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   autogen
    #   pyautogen
pytz==2024.2
    # via pandas
pywin32==306
    # via
    #   docker
    #   portalocker
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   azure-core
    #   docker
    #   msal
    #   streamlit
    #   tiktoken
rich==13.8.1
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
streamlit==1.35.0
streamlit_extras
    # via -r requirements.in
tenacity==8.5.0
    # via streamlit
termcolor==2.4.0
    # via
    #   autogen
    #   pyautogen
tiktoken==0.7.0
    # via
    #   autogen
    #   pyautogen
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   altair
    #   azure-core
    #   azure-identity
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.3
    # via
    #   docker
    #   requests
watchdog==5.0.2
    # via streamlit